Modify: dlOS.sh

+++ dlOS.sh
@@ -4,7 +4,8 @@ FILENAME=$1
 REPOURL="https://raw.githubusercontent.com/verynerd/macOS/master/"
 CDNREPO="https://cdn.jsdelivr.net/gh/verynerd/macOS@resources/"
 
-PS3='Specify the version of macOS you want:'
+PS3='Enter the numerical number of the option: '
+echo "Specify the version of macOS you want."
 options=("macOS 10.14.6" "macOS 10.15.6" "macOS 10.16 Beta 7" "Quit")
 select opt in "${options[@]}"
 do
@@ -21,7 +22,6 @@ do
             ;;
         "macOS 10.16 Beta 7")
             SYSVER="10_16_Beta7"
-            SYSVER="10_16_Beta7"
             URL="${REPOURL}${SYSVER}_links.txt"
             break
             ;;
@@ -29,7 +29,7 @@ do
             echo ""    
             break
             ;;
-        *) echo "Invalied input";;
+        *) echo "Invalied input! Please enter the numerical number of the option";;
     esac
 done
 
@@ -56,21 +56,26 @@ do
     esac
 done
 
+# Exit if SYSVER NOT SPECIFIED
+if [[ $SYSVER != '' ]]; then
 
-if [[ -d ~/Downloads/macOS ]]; then
+    if [[ -d ~/Downloads/macOS ]]; then
 
-	rm -r ~/Downloads/macOS
-	mkdir ~/Downloads/macOS && cd  ~/Downloads/macOS
+        rm -r ~/Downloads/macOS
+        mkdir ~/Downloads/macOS && cd  ~/Downloads/macOS
 
-else
-	mkdir ~/Downloads/macOS && cd  ~/Downloads/macOS
+    else
+        mkdir ~/Downloads/macOS && cd  ~/Downloads/macOS
+    fi
+
+    # 下载 plist 文件
+    wget -q --show-progress --no-check-certificate -O links.txt $URL 
+
+    echo "Downloading split files..."
+    mkdir splits && wget -q --show-progress --no-check-certificate -i links.txt -P splits 
+
+    cat splits/${SYSVER}.dmg* > ${SYSVER}.dmg 
+    rm -R splits    
 fi
 
-# 下载 plist 文件
-wget -q --show-progress --no-check-certificate -O links.txt $URL 
 
-echo "Downloading split files..."
-mkdir splits && wget -q --show-progress --no-check-certificate -i links.txt -P splits 
-
-cat splits/${SYSVER}.dmg* > ${SYSVER}.dmg 
-rm -R splits
